fix step 2
