+++ public/index.html
@@ -29,6 +29,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <div id="modals"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
