+++ src/components/Modal/Modal.module.css
@@ -1,0 +1,35 @@
+.modalOverlay {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 1000;
+}
+
+.modal {
+    position: relative;
+    background-color: #1C1C21;
+    padding: 30px;
+    border-radius: 25px;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.2);
+    width: 600px;
+    height: 530px;
+    z-index: 1010;
+}
+
+.modalHeader {
+    display: flex;
+    justify-content: flex-end;
+}
+
+.closeBtn {
+    background: none;
+    border: none;
+    cursor: pointer;
+}
